new req for updateds

+++ updateds_kai.py
@@ -13,6 +13,7 @@ import datetime as datetime
 from datetime import date
 from xml.etree import ElementTree as ET
 from xml.etree.ElementTree import fromstring
+from xml.etree.ElementTree import ParseError
 import os
 import random
 from sqlalchemy import create_engine
@@ -45,10 +46,24 @@ with open(os.path.join(os.path.dirname(os.path.abspath(__file__)), 'secrets.json
 
 # ['D', 'IS', 'A', 'SF', 'I', 'EI', 'NL', 'CH', 'GB' ]
 
+TIME_OUT = 10
+
+def clean_min_pax(set_data):
+	if 1 in set_data:
+		set_data.clear()
+		set_data.add(1)
+
+def clean_min_pax_child(l):
+	for child in l:
+		if child['min_pax'] == 1:
+			pass
+			
+
 @click.command()
-@click.option('--country', default='IA')
+# @click.option('--country', default='IA')
+@click.option('--req/--no-req', default=True)
 # @click.option('--client', default='ctrip')
-def updateds(country):
+def updateds(req):
 
 	url = 'https://rbs.gta-travel.com/rbscnapi/RequestListenerServlet'
 	res = []
@@ -65,7 +80,7 @@ def updateds(country):
 	engine = create_engine('sqlite:///' + os.path.join(os.path.dirname(os.path.abspath(__file__)), 'destServ.db'))
 	# 
 	# services_raw = engine.execute("SELECT * FROM destination_service_raw WHERE country_code='{0}';".format(country))
-	services_raw = engine.execute("SELECT * FROM destination_service_raw;".format(country))
+	services_raw = engine.execute("SELECT * FROM destination_service_raw;")
 
 	n_d = datetime.datetime.now().date()
 
@@ -86,10 +101,11 @@ def updateds(country):
 		# if client == 'ctrip':
 		# 	if not ('#'.join([row['city_code'], row['item_code']]) in reqs):
 		# 		continue
-		gta_key = '_'.join([row['city_code'], row['item_code']])
-		if not gta_key in reqs:
-			# print('Not in top selling .. skipping .. ' + str(gta_key))
-			continue
+		if req:
+			gta_key = '_'.join([row['city_code'], row['item_code']])
+			if not gta_key in reqs:
+				# print('Not in top selling .. skipping .. ' + str(gta_key))
+				continue
 
 		if not (last_city_code == row['city_code'] and last_item_code == row['item_code']):
 			entry = {}
@@ -104,6 +120,7 @@ def updateds(country):
 		min_pax = int(row['min_pax'])
 		if row['pax_type'] == 'Adult' and min_pax <= 9:
 			entry['adult_min_pax'].add(min_pax)
+			# clean_min_pax(entry['adult_min_pax'])
 		elif row['pax_type'] == 'Child' and min_pax <= 9:
 			child = {}
 			child['age_from'] = row['age_from']
@@ -137,12 +154,16 @@ def updateds(country):
 		print('Fetching static info.. ' + str(counter) + ': ' + '_'.join([service['city_code'], service['item_code']]))
 
 		try:
-			ri = requests.post(url, data=ET.tostring(si_tree.getroot(), encoding='UTF-8', method='xml'), timeout=360)
+			ri = requests.post(url, data=ET.tostring(si_tree.getroot(), encoding='UTF-8', method='xml'), timeout=TIME_OUT)
 		except OSError:
 			pprint.pprint('Error: ignoring OSError...')
 			continue
 
-		ri_tree = ET.fromstring(ri.text)
+		try:
+			ri_tree = ET.fromstring(ri.text)
+		except ParseError:
+			print('Error: Parser error.. i')
+			continue
 
 		if ri_tree.find('.//Errors') != None:
 			if not len(list(ri_tree.find('.//Errors'))):
@@ -322,7 +343,8 @@ def updateds(country):
 					# pprint.pprint(ET.tostring(sp_tree.getroot(), encoding='UTF-8', method='xml'))
 
 					try:
-						rp = requests.post(url, data=ET.tostring(sp_tree.getroot(), encoding='UTF-8', method='xml'), timeout=600)
+						# rp = requests.post(url, data=ET.tostring(sp_tree.getroot(), encoding='UTF-8', method='xml'), timeout=600)
+						rp = requests.post(cur_clients[client]['url'], data=ET.tostring(sp_tree.getroot(), encoding='UTF-8', method='xml'), timeout=TIME_OUT)
 					except OSError:
 						pprint.pprint('Error: ignoring OSError...')
 						continue
@@ -341,8 +363,13 @@ def updateds(country):
 						continue
 
 					# pprint.pprint(rp.text)
+					# service['currency'] = 'Unkown'
 
-					if rp_tree.find('.//SightseeingDetails') == None or not len(list(rp_tree.find('.//SightseeingDetails'))):
+					if rp_tree.find('.//SightseeingDetails') == None:
+						print('Warning: no sightseeing ele.. i')
+						continue
+
+					if not len(list(rp_tree.find('.//SightseeingDetails'))):
 						pprint.pprint('No sightseeing price returned...')
 						continue
 
@@ -420,7 +447,9 @@ def updateds(country):
 					# pprint.pprint(ET.tostring(sp_tree.getroot(), encoding='UTF-8', method='xml'))
 
 					try:
-						rp = requests.post(url, data=ET.tostring(sp_tree.getroot(), encoding='UTF-8', method='xml'), timeout=600)
+						# rp = requests.post(url, data=ET.tostring(sp_tree.getroot(), encoding='UTF-8', method='xml'), timeout=600)
+						rp = requests.post(cur_clients[client]['url'], data=ET.tostring(sp_tree.getroot(), encoding='UTF-8', method='xml'), timeout=TIME_OUT)
+						# cur_clients[client]['url']
 					except OSError:
 						pprint.pprint('Error: ignoring OSError...')
 						continue
@@ -437,6 +466,10 @@ def updateds(country):
 						continue
 
 					# pprint.pprint(ET.tostring(rp_tree.getroot(), encoding='UTF-8', method='xml'))
+
+					if rp_tree.find('.//SightseeingDetails') == None:
+						print('Warning sightseeing detials is none.. child')
+						continue
 
 					if not len(list(rp_tree.find('.//SightseeingDetails'))):
 						pprint.pprint('No child sightseeing price returned...')
@@ -482,27 +515,42 @@ def updateds(country):
 					# service['tour_operations'].append(tour_operation)
 					service[mapped_tour_key].append(tour_operation)
 
+			# for multi client, bad hack..
 			if counter_client == 0:
 				services_p.append(service)
+			if (counter+1) % 10 == 0:
+				columns = 'country, city_code, item_code, name, duration, summary, please_note, includes, the_tour, additional_information, currency, policy, tour_operations, tour_operations1, tour_operations2, tour_operations3, tour_operations4, tour_operations5, tour_operations6, tour_operations7, closed_dates, thumb_nail, image'
 
-	columns = 'country, city_code, item_code, name, duration, summary, please_note, includes, the_tour, additional_information, currency, policy, tour_operations, tour_operations1, tour_operations2, tour_operations3, tour_operations4, tour_operations5, tour_operations6, tour_operations7, closed_dates, thumb_nail, image'
+				print('INFO: Inserting into db.. ' + str(counter))
+				for service in services_p:
+					engine.execute("DELETE FROM destination_service WHERE city_code='{0}' AND item_code='{1}';".format(service['city_code'], service['item_code']))
+				
+				# print(str(services_p))
+				for service in services_p:
+					# print('INFO: Inserting.. ' + str(service))
+					print('INFO: Inserting.. ')
+					if len(service['tour_operations']) != 0:
+						if 'currency' not in service.keys():
+							service['currency'] = 'unknown'
 
-	for service in services_p:
-		engine.execute("DELETE FROM destination_service WHERE city_code='{0}' AND item_code='{1}';".format(service['city_code'], service['item_code']))
-	for service in services_p:
-		if len(service['tour_operations']) != 0:
-			r = [service['country'], service['city_code'], service['item_code'], \
-				service['name'], service['duration'], service['summary'], \
-				service['please_note'], service['includes'], service['the_tour'], service['additional_information'], \
-				service['currency'], json.dumps(service['policy']), json.dumps(service['tour_operations']), json.dumps(service['tour_operations1']), json.dumps(service['tour_operations2']), json.dumps(service['tour_operations3']), json.dumps(service['tour_operations4']), json.dumps(service['tour_operations5']), json.dumps(service['tour_operations6']), json.dumps(service['tour_operations7']), \
-				json.dumps(service['closed_dates']), service['thumb_nail'], service['image']]
-				# with engine.connect() as connection:
-					# connection.execute(text("INSERT INTO destination_service ({0}) VALUES({1});".format(columns, ','.join( '\'' + ent.replace('\'', '\'\'') + '\'' for ent in r))).execution_options(autocommit=True) )
-					# connection.commit()
-			if r == None:
-				print('Error: r is none when inserting into database..')
-				continue
-			engine.execute("INSERT INTO destination_service ({0}) VALUES({1});".format(columns, ','.join( '\'' + ent.replace('\'', '\'\'') + '\'' for ent in r) ))
+						for col in columns:
+							if col not in service.keys():
+								service[col] = 'unknown'
+
+						r = [service['country'], service['city_code'], service['item_code'], \
+							service['name'], service['duration'], service['summary'], \
+							service['please_note'], service['includes'], service['the_tour'], service['additional_information'], \
+							service['currency'], json.dumps(service['policy']), json.dumps(service['tour_operations']), json.dumps(service['tour_operations1']), json.dumps(service['tour_operations2']), json.dumps(service['tour_operations3']), json.dumps(service['tour_operations4']), json.dumps(service['tour_operations5']), json.dumps(service['tour_operations6']), json.dumps(service['tour_operations7']), \
+							json.dumps(service['closed_dates']), service['thumb_nail'], service['image']]
+							# with engine.connect() as connection:
+								# connection.execute(text("INSERT INTO destination_service ({0}) VALUES({1});".format(columns, ','.join( '\'' + ent.replace('\'', '\'\'') + '\'' for ent in r))).execution_options(autocommit=True) )
+								# connection.commit()
+						if r == None:
+							print('Error: r is none when inserting into database..')
+							continue
+						engine.execute("INSERT INTO destination_service ({0}) VALUES({1});".format(columns, ','.join( '\'' + ent.replace('\'', '\'\'') + '\'' for ent in r) ))
+						# print("INSERT INTO destination_service ({0}) VALUES({1});".format(columns, ','.join( '\'' + ent.replace('\'', '\'\'') + '\'' for ent in r) ))
+				services_p.clear()
 
 if __name__ == '__main__':
 	updateds()
